% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getGooglePrice.R
\name{getGooglePrice}
\alias{getGooglePrice}
\title{Get prices from google including intraday}
\usage{
getGooglePrice(symbol, freq = 60, period = "1d", exchng = "SGX",
  tz = "Asia/Singapore")
}
\value{
a data table of all the financials data
}
\description{
Get prices from google including intraday
}

