% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getSGXStocks.R
\name{getSGXStocks}
\alias{getSGXStocks}
\title{Get list of all stocks in SGX
Returns a datatable}
\usage{
getSGXStocks()
}
\value{
a datatable of all SGX stocks
}
\description{
Get list of all stocks in SGX
Returns a datatable
}

