% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getFinancials.R
\name{getWSJFinancials}
\alias{getWSJFinancials}
\title{Get prices from yahoo}
\usage{
getWSJFinancials(ticker, interval = "quarter")
}
\value{
a data table of all the financials data
}
\description{
Get prices from yahoo
}

