% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getYahooPrice.R
\name{getYahooPrice}
\alias{getYahooPrice}
\title{Get prices from yahoo}
\usage{
getYahooPrice(env, Symbols, from = "2007-01-01", to = Sys.Date(), ...)
}
\value{
a environment . to access , e.g., getYahooPrice$INTC
}
\description{
Get prices from yahoo
}

